requirements.txt
bcrypt==4.0.1
blinker==1.7.0
click==8.1.7
colorama==0.4.6
Flask==3.0.0
Flask-MySQLdb==2.0.0
importlib-metadata==6.8.0
itsdangerous==2.1.2
Jinja2==3.1.2
MarkupSafe==2.1.3
mysql-connector-python==8.2.0
mysqlclient==2.2.0
numpy==1.26.2
pandas==2.1.4
password-strength==0.0.3.post2
protobuf==4.21.12
python-dateutil==2.8.2
pytz==2023.3.post1
six==1.16.0
tzdata==2023.3
Werkzeug==3.0.1
zipp==3.17.0